+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - VMware Plugin: Adapt to Python 3.12 [PR #1879]
 - freebsd: fix build issues with ports tree 2024Q3 [PR #1884]
 - matrix remove obsolete SUSE [PR #1906]
+- VMware Plugin: Add option restore_allow_disks_mismatch [PR #1905]
 
 ### Fixed
 - fix sql error on bad virtualfull; detect parsing errors with strtod [PR #1842]
@@ -508,5 +509,6 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1886]: https://github.com/bareos/bareos/pull/1886
 [PR #1890]: https://github.com/bareos/bareos/pull/1890
 [PR #1894]: https://github.com/bareos/bareos/pull/1894
+[PR #1905]: https://github.com/bareos/bareos/pull/1905
 [PR #1906]: https://github.com/bareos/bareos/pull/1906
 [unreleased]: https://github.com/bareos/bareos/tree/master